WIP Starting down the path of doing Unobtrusive Java Script the Rails way, I realised that this Rails-AJAX pattern isn't actually designed to work with ActionCable.
The problem is that Rails would like me to put my js about updating the page in the views/.../*.js.erb file
but I need the updates to also be performed by the other clients (the other players).

The good thing about updating the page via the *.js.erb file is that it allows me to slip into Ruby in
order to render a partial. There won't be a way to render partials from vanilla JS (which is what the
other clients will need to be using).

However, I don't see a way around having to use vanilla JS, so I may as well do that for everyone
rather than render a partial for the non-other player.

But now this means I am duplicating the partial code: besides the partial, for server-side rendering,
I am writing it in JS, too, for client-side rendering.