Use AJAX for game start button
data-remote=true is the line that does this.

Unless we do this, the page will attempt to reload (which happens to do nothing because
we don't have a template for GamesController:start)
This is an issue because it unsubscribes the player from the game stream temporarily
so that they don't receive the ActionCable
message containing the info needed to display
the game start modal dialog.